% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter-methods.R
\docType{methods}
\name{getter}
\alias{getter}
\alias{$,mdIVW-method}
\title{Applies method $ to different classes}
\usage{
\S4method{$}{mdIVW}(x, name)
}
\arguments{
\item{x}{Object.}

\item{name}{Name of slot.}
}
\description{
Enables slots of objects in this package to be referenced easily.
}
\keyword{internal}
