% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{mdIVW-class}
\alias{mdIVW-class}
\title{PIVW Class}
\description{
An object containing the estimate produced using the penalized inverse-variance weighted (pIVW) method as well as various statistics.
}
\section{Slots}{

\describe{
\item{\code{Over.dispersion}}{Should the method consider overdispersion (balanced horizontal pleiotropy)? Default is TRUE.}

\item{\code{Delta}}{The z-score threshold for IV selection. By default, \code{delta=0} (i.e., no IV selection will be conducted).}

\item{\code{Estimate}}{The causal point estimate from the pIVW estimator.}

\item{\code{StdError}}{The standard error associated with \code{Estimate}.}

\item{\code{CILower}}{The lower bound of the confidence interval for \code{Estimate}, which is derived from the bootstrapping Fieller method or normal distribution. For the bootstrapping Fieller's interval, if it contains multiple ranges, then lower limits of all ranges will be reported.}

\item{\code{CIUpper}}{The upper bound of the confidence interval for \code{Estimate}, which is derived from the bootstrapping Fieller method or normal distribution. For the bootstrapping Fieller's interval, if it contains multiple ranges, then upper limits of all ranges will be reported.}

\item{\code{Alpha}}{The significance level used in constructing the confidence interval (default is 0.05).}

\item{\code{Pvalue}}{P-value associated with the causal estimate from the pIVW estimator.}

\item{\code{Tau2}}{The variance of the balanced horizontal pleiotropy. \code{Tau2} is calculated by using all IVs in the data before conducting the IV selection.}

\item{\code{SNPs}}{The number of SNPs after IV selection.}

\item{\code{Condition}}{The estimated effective sample size. It is recommended to be greater than 5 for the pIVW estimator to achieve reliable asymptotic properties.}
}}

\keyword{internal}
